refactor: 抽离 InboundEventParser（E2 Step 1b）
